refactor(backend): remove useless modules
