+++ apps/backend/src/migrator/m20230509_000008_create_media_import_report.rs
@@ -3,7 +3,7 @@ use sea_orm::{DeriveActiveEnum, EnumIter};
 use sea_orm_migration::prelude::*;
 use serde::{Deserialize, Serialize};
 
-use super::m20230417_000004_create_user::User;
+use super::m20230417_000002_create_user::User;
 
 pub struct Migration;
